distribute bragi tests in files depending on data used (no more reload)
